Multiple UX improvments and bug fixes
Added back button functionality that will allow users to return to
their schedules without needing to log back in

Fixed some app crashes that were happening when users decide to spam
tap a UITableViewCell in the Classes Tab

Added improved checks for cases where there is no internet and/or users
aren’t connected to facebook. If they arent connected to fb, they will
be given a chance to login to facebook.